+++ src/common/processing.h
@@ -8,10 +8,7 @@ static inline const char* ffProcessAppendStdOut(FFstrbuf* buffer, char* const ar
 {
     const char* error = ffProcessAppendOutput(buffer, argv, false);
     if (!error)
-    {
-        ffStrbufTrimRight(buffer, '\n');
-        ffStrbufTrimRight(buffer, ' ');
-    }
+        ffStrbufTrimRightSpace(buffer);
     return error;
 }
 
@@ -19,9 +16,6 @@ static inline const char* ffProcessAppendStdErr(FFstrbuf* buffer, char* const ar
 {
     const char* error = ffProcessAppendOutput(buffer, argv, true);
     if (!error)
-    {
-        ffStrbufTrimRight(buffer, '\n');
-        ffStrbufTrimRight(buffer, ' ');
-    }
+        ffStrbufTrimRightSpace(buffer);
     return error;
 }