+++ src/logo/builtin.c
@@ -209,6 +209,7 @@ static const FFlogo A[] = {
     // AoscOsRetro_small
     {
         .names = {"Aosc OS/Retro_small", "aoscosretro_small"},
+        .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
         .lines = FASTFETCH_DATATEXT_LOGO_AOSCOSRETRO_SMALL,
         .colors = {
             FF_COLOR_FG_BLUE,
@@ -764,6 +765,7 @@ static const FFlogo C[] = {
     // CalinixOSSmall
     {
         .names = {"Calinix_small", "calinixos_small"},
+        .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
         .lines = FASTFETCH_DATATEXT_LOGO_CALINIXOS_SMALL,
         .colors = {
             FF_COLOR_FG_MAGENTA,
@@ -875,7 +877,7 @@ static const FFlogo C[] = {
     // Chimera Linux
     {
         .names = {"Chimera Linux"},
-        .lines = FASTFETCH_DATATEXT_LOGO_CHIMERALINUX,
+        .lines = FASTFETCH_DATATEXT_LOGO_CHIMERA_LINUX,
         .colors = {
             FF_COLOR_FG_RED,
             FF_COLOR_FG_MAGENTA,
@@ -1241,9 +1243,9 @@ static const FFlogo D[] = {
         .colorKeys = FF_COLOR_FG_RED,
         .colorTitle = FF_COLOR_FG_WHITE,
     },
-    // Drauger
+    // DraugerOS
     {
-        .names = {"Drauger"},
+        .names = {"DraugerOS", "Drauger"},
         .lines = FASTFETCH_DATATEXT_LOGO_DRAUGER,
         .colors = {
             FF_COLOR_FG_RED,
@@ -1336,8 +1338,8 @@ static const FFlogo E[] = {
     // EndeavourSmall
     {
         .names = {"Endeavour_small", "endeavour-linux_small", "endeavouros_small", "endeavouros-linux_small"},
-        .lines = FASTFETCH_DATATEXT_LOGO_ENDEAVOUR_SMALL,
         .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
+        .lines = FASTFETCH_DATATEXT_LOGO_ENDEAVOUR_SMALL,
         .colors = {
             FF_COLOR_FG_RED,
             FF_COLOR_FG_MAGENTA,
@@ -1823,6 +1825,7 @@ static const FFlogo G[] = {
     // GuixSmall
     {
         .names = {"Guix_small"},
+        .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
         .lines = FASTFETCH_DATATEXT_LOGO_GUIX_SMALL,
         .colors = {
             FF_COLOR_FG_YELLOW,
@@ -2088,6 +2091,15 @@ static const FFlogo K[] = {
         },
         .colorKeys = FF_COLOR_FG_BLUE,
         .colorTitle = FF_COLOR_FG_WHITE,
+    },
+    // KernelOS
+    {
+        .names = {"KernelOS"},
+        .lines = FASTFETCH_DATATEXT_LOGO_KERNELOS,
+        .colors = {
+            FF_COLOR_FG_RED,
+            FF_COLOR_FG_MAGENTA,
+        }
     },
     // KDENeon
     {
@@ -2462,6 +2474,7 @@ static const FFlogo M[] = {
     // MageiaSmall
     {
         .names = {"Mageia_small"},
+        .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
         .lines = FASTFETCH_DATATEXT_LOGO_MAGEIA_SMALL,
         .colors = {
             FF_COLOR_FG_CYAN,
@@ -2649,7 +2662,7 @@ static const FFlogo M[] = {
     },
     // MX
     {
-        .names = {"MX"},
+        .names = {"MX", "MX Linux"},
         .lines = FASTFETCH_DATATEXT_LOGO_MX,
         .colors = {
             FF_COLOR_FG_WHITE,
@@ -2659,8 +2672,20 @@ static const FFlogo M[] = {
     },
     // MXSmall
     {
-        .names = {"MX_small", "mx-small"},
+        .names = {"MX_small", "mx linux_small"},
+        .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
         .lines = FASTFETCH_DATATEXT_LOGO_MX_SMALL,
+        .colors = {
+            FF_COLOR_FG_WHITE,
+        },
+        .colorKeys = FF_COLOR_FG_BLUE,
+        .colorTitle = FF_COLOR_FG_CYAN,
+    },
+    // MX2
+    {
+        .names = {"MX2"},
+        .type = FF_LOGO_LINE_TYPE_ALTER_BIT,
+        .lines = FASTFETCH_DATATEXT_LOGO_MX2,
         .colors = {
             FF_COLOR_FG_WHITE,
         },
@@ -2730,6 +2755,7 @@ static const FFlogo N[] = {
     // NixOSSmall
     {
         .names = {"NixOS_small", "nix_small", "nixos-linux-small", "nix-linux-small"},
+        .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
         .lines = FASTFETCH_DATATEXT_LOGO_NIXOS_SMALL,
         .colors = {
             FF_COLOR_FG_BLUE,
@@ -3301,6 +3327,7 @@ static const FFlogo P[] = {
     // PostMarketOSSmall
     {
         .names = {"PostMarketOS_small"},
+        .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
         .lines = FASTFETCH_DATATEXT_LOGO_POSTMARKETOS_SMALL,
         .colors = {
             FF_COLOR_FG_GREEN,
@@ -3560,6 +3587,7 @@ static const FFlogo R[] = {
     // RockyLinuxSmall
     {
         .names = {"rocky_small", "rocky-linux_small", "rockylinux_small"},
+        .type = FF_LOGO_LINE_TYPE_SMALL_BIT,
         .lines = FASTFETCH_DATATEXT_LOGO_ROCKY_SMALL,
         .colors = {
             FF_COLOR_FG_GREEN,